[rename]vueファイル名をリネーム
