+++ vite.config.js
@@ -3,6 +3,9 @@ import laravel from "laravel-vite-plugin";
 import vue from "@vitejs/plugin-vue";
 
 export default defineConfig({
+    resolve: {
+        preserveSymlinks: true, // シンボリックリンクを辿っていないパスでファイルの同一性を判別 ※ファイル名をリネームした時に役立つ
+    },
     plugins: [
         vue(),
         laravel({
